Checkstyle configuration to not scan generated sources
I've noticed "red" CS errors in Eclipse using
https://github.com/m2e-code-quality/m2e-code-quality in some of the AAA
projects, and upon closer inspection, found that it complained about
generated code, which I think is useful (and the less stuff we don't
have to we don't scan, the less impact on the build time our quality
checks have).

Also includes clean-up moves to make the <configuration> easier to
understand.  This should, hopefully, not change whata is scanned.

Change-Id: Icbca680774fa553b03554186efa88c72349a2df9
Signed-off-by: Michael Vorburger <vorburger@redhat.com>