Checkstyle <excludes> on single line and plus target[-ide]
Adding target & target-ide to <excludes> .. makes no difference yet, but
done in anticpation of a future change to scan full project, based on an
idea I've found in aaa's Checkstyle configuration.

Also clean-up of something, as I'm not convinced the multi-line
<excludes> as is actually ever worked.. this Maven property is not a
List, but a String; output from mvn -X gives reason to believe that this
didn't work as intended.

Change-Id: Ied595a47c322d9ac6b5734edeccc19273da0f699
Signed-off-by: Michael Vorburger <vorburger@redhat.com>