Update patch set 2
Patch Set 2:

I think 2.4.13 will be out quite quickly. Other option would be to re-package it in odlparent and publish it as a -SNAPSHOT -- that way we can remove it as soon as it is no longer necessary. I think waiting a few weeks is okay -- scala bump is blocked on karaf anyway.

Patch-set: 2