Immutables.org in dependencyManagement
https://immutables.github.io is really great IMHO.  So far I've used it
to write a few structs for test data, have made some small contributions
to it (bug reports & testing; no code) but I'd like to start using it
more in ODL work, and thus am proposing it into odlparent.  It's a fast
moving project, so having a single place (odlparent) for the version and
keep moving it forward is useful from the start IMHO, instead of
plastering it across projects.

I do realize this code. gen. could be seen under the YANG lens, but IMHO
it's complementary and has a different sweet spot (YANG for persistent
data store, remotable objects) and this Immutables.org for private
in-bundle code for tests and configuration data - notably anything
threading related such as https://git.opendaylight.org/gerrit/#/c/48373/
discussion, where easy true immutability is of interest (YANG gen. code
is NOT immutable; it pretends to be, but it's actually not at all).

Change-Id: I1bc87c3d11316aa42002eb1a4b1f49f820058860
Signed-off-by: Michael Vorburger <vorburger@redhat.com>