Add buildnumber Maven plugin
This allows projects to obtain a ${buildNumber} property containing
the git hash of the commit being built, simply by activating the
buildnumber-maven-plugin. (It can't be activated by default, because
it needs an SCM URL even if it just pulls in the local git hash.)

See
https://blog.jayway.com/2012/04/07/continuous-deployment-versioning-and-git/
for examples. We could use this to include the git hash in our
bundles' manifests and in our distribution archives. I'm suggesting
adding this to odlparent-lite rather than odlparent because we may
want to identify non-code project artifacts!

This came out of a discussion with Josh Hershberg.

Change-Id: I0b3fc9a7a956499ff6fdff5fbbfb32361de74f53
Signed-off-by: Stephen Kitt <skitt@redhat.com>