git-commit-id-plugin to put a META-INF/git.properties in all built JAR
see https://github.com/ktoso/maven-git-commit-id-plugin

Contrary to another similar proposed change which is based on the
buildnumber-maven-plugin that has a similar purposes, this one
automatically works for all ODL JARs without any changes needed in any
other POM (because it does not rely on a Maven SCM URL).

The git.properties is ACTUALLY not yet in all built JAR, only in non
OSGi bundle JARs (such as some *-testutils), because there is some issue
probably with the maven-bundle-plugin (?); this isn't related to this,
but also e.g. the LICENSE file in target/classes doesn't yet make it
into OSGi bundle JARs; I'll try to address that with a separate Gerrit
next.

Change-Id: I43f9692b4a3ae83a90cdfbeaf20bd2cba323bce8
Signed-off-by: Michael Vorburger <vorburger@redhat.com>