Fix broken build of inject-guice-testutils
by making Guice a default <scope>compile instead of a <scope>test.  The
problem with <scope>test, due to the way Maven works, is that in a
project like inject-guice-testutils which has a <dependency> to Mycila
with <scope>compile, Guice will still be forced to be <scope>test, which
breaks the build of inject-guice-testutils (because its src/main/java
code depends on Guice).  

While this could of course be solved by repeating Guice in the
inject-guice-testutils POM and force it too to be <scope>compile, this
is very confusing, and it's thus perhaps best to not overuse <scope> in
<dependencyManagement> and let using projects decide the scope.

Also, and technically unrelated, swapping places of recently merged
Guice and Immutables, just so that Guice is right after Mycila as
intended; I mixed that up during the rebase earlier today.

Change-Id: Ib6f37b0f45f8787c95b498ea56a16fc954af804a
Signed-off-by: Michael Vorburger <vorburger@redhat.com>