Update patch set 1
Patch Set 1:

> This suggests a dependency and/or scope issue on the
 > inject-guice-testutils side of things. In odlparent we specify the
 > scope that we expect dependencies to be used in, which is test for
 > both Mycila and Guice. If inject-guice-testutils needs another
 > scope, it needs to redefine it appropriately — on both dependencies
 > if necessary.

OK I'll put in another Gerrit to address it on that end then.

 > Another solution of course would be to make inject-guice-testutils
 > a scope-test dependency too (producing a test-jar).

The problem with the test-jar approach, if we mean the same thing, is that it doesn't propagate transitive dependencies.  This is why I'm going for the *-testutils approach without attached test-jar.

Patch-set: 1
Reviewer: Gerrit User 5867 <5867@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0