Update patch set 2
Patch Set 2:

> OK, great - then I've restored this one!  Of course, we could also
 > simply change Sonar rules... ;-) Do you, personally, think this is
 > a useful rule?  Or do you just want to not have Sonar issue, but
 > actually think this is a stupid rule?  I'm curious, because I'm
 > leaning towards the latter.

Coming from a mostly C background, I'm quite used to have short trailing comments, and while I see the points described as the reason behind this rule in Sonar, I don't find it a particularly useful rule, I just wanted the issues gone, and not being introduced again in the future. So I'm leaning towards the latter too, but I never asked if we can change those rules.

 > 
 > Will you +1 this when you've had a chance to getting all other
 > projects to be compliant too?  FYI https://wiki.opendaylight.org/view/EvolvingCheckstyle
 > has steps how I do this, hopefully useful?  I'm just applying that
 > for https://git.opendaylight.org/gerrit/#/c/43331/, which is much
 > higher value than this one, to me..

If we can disable the Sonar rule for this issue, then no need, if we don't, I will.

Patch-set: 2