Update patch set 1
Patch Set 1: Code-Review+1

This makes sense in a couple of ways: it ensures we run SFT with the same Karaf patches as we use for the distribution (Aries, Xerces), and it ensures we run SFT with the same boot features as we use in the distribution.

Do we know if this is liable to cause SFT failures in downstream projects?

Patch-set: 1
Reviewer: Gerrit User 4540 <4540@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1