Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Well, lifecycle mapping should be defined where that execution is triggered. That is to say: if model-parent is the first thing to define an execution for yang-maven-plugin, then the lifecycle mapping should be defined exactly there.

If I open a pom.xml, I do not expect to see mappings for executions which are not defined there, embarking on either a chase through parents or (shudder) examining the effective model.

And that brings us to the oomph-vs-pom declaration of lifecycle mapping :)

Patch-set: 1
Reviewer: Gerrit User 649 <649@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1