Update patch set 1
Patch Set 1:

This is not going to be a smooth ride, as FutureFallback was removed in Guava-20 and the corresponding change in Futures has not been introduced until Guava-19.

Downstream fix in yangtools will be quite simple, so if we coordinate well, we should not be seeing more than one day of breakage.

Patch-set: 1
Label: Verified=0