Update patch set 1
Patch Set 1:

Looks good to me in principle on reviewing it here, but let me try out pulling and running with this under Eclipse, to confirm it's all fine there too, before you merge it?  I want to make sure that its Source Folders are still correct after this.

Patch-set: 1