Fix karaf's aries-blueprint feature
quiesce.api is something we are using and since blueprint-core has
it as an optional dependency, we end up refreshing blueprint when
we install it. That leads to pax-exam thinking the test during SFT
has ended and proceeds to the next test -- wreaking havoc in process
of doing so.

Patch the upstream feature file so quiesce.api is installed
unconditionally, so we do not have to install it from a feature,
hence we won't trigger blueprint refresh.

Change-Id: I14ad8b70d1dc8a961944da9087a54242ba017195
Signed-off-by: Robert Varga <rovarga@cisco.com>