bundle-parent maven-bundle-plugin instructions include bnd.bnd
This allows BND instructions to be in a file named bnd.bnd next to the
pom.xml, instead of inline in the pom.xml.  This keeps pom.xml files
shorter, and make BND instructions useable by tools such as the Eclipse
plugin bndtools, or a future alternative build system based e.g. on
Gradle instead of a Maven.

Change-Id: Ifba1cf3f0137003a786575acc862e9e9c0f711fa
Signed-off-by: Michael Vorburger <vorburger@redhat.com>