Update patch set 3
Patch Set 3:

> IMHO, the <name> should be used for human consumption in mind,
 > having a user friendly short description of the Maven project, and
 > leaving computer readable stuff to other places.

Most projects have the default <name> field configured which is <name>${project.artifactId}</name> so I think this is a slight improvement.

The reason we use <name> is because it's what affects the console output which is what we are using to parse out the build failure, without it we won't be able to determine which project is causing build failure. Although the only feature we're using the <name> field for is to automatically email the ${project}-dev mailing list.

Worst case if we cannot parse out the ${project} we are still emailing the release mailing list so it would just involve an additional step of a human determining which project to cc after the fact.

Patch-set: 3