Update patch set 1
Patch Set 1:

I think feature-repo-parent would still add one feature (named the same as the repo). I do not remember if there is a Bug open for not creating that feature, but that can be fixed in a separate change anyway.

Otherwise looks good, but it obviously should be tested to make sure no project breaks.

Patch-set: 1