Bug 8175: Improve SFT TestBundleDiag under Karaf 4
This does NOT yet completely fix all problems of the SingleFeatureTest
TestBundleDiag under Karaf 4 (which was in really bad shape!), most
importantly there is still something somewhere in Pax Exam that I
haven't understood yet which seems to just "swallow" all exceptions
thrown by SingleFeatureTest's @Test installFeature() but at least now:

1. there is something in the log now for failures at least
   thanks to a (try/catch/log) WORKAROUND (!)

2. LinkageError fixed (see details below)

3. IllegalStateException: BundleContext is no longer valid fixed (?)

This changes the setup by just embedding bundles4-test and it's
dependencies into features4-test, instead of programmatically installing
the odl-bundles-test feature (which is removed in this change) or the
features4-test bundle.  

This simplifies installation, avoid problems with the wrap: protocol,
and should resolve weird problems we've seen (only!) in the distribution
jobs, which seem to have had something to do with a custom local Maven
repo.

IMHO it would be *REALLY* good to practice proper TDD for SFT (with
TestBundleDiag) now, and have some automated negative testing of broken
features and bundles (incl. BP) in odlparent...

Bug: 7981 related later proper solution
Change-Id: I516c0f6c4e9ee85f0c753cb4b58a0b53fc22c7ea
Signed-off-by: Michael Vorburger <vorburger@redhat.com>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>