Update patch set 4
Patch Set 4:

> I cannot reproduce this locally

Looking at the log, this is happening in step five of distribution-check (as opposed to step three).
See the definition of distribution-check in releng/builder: jjb/releng-templates-java.yaml 

The tell are the lines with "Downloaded: file:///tmp/n" (as opposed to downloading from Nexus).
With Bug 7974, this is a second instance of a mysterious failure which only happens in step 5 (and not in step 3 already). I am currently working on updating distribution-check, perhaps when I am done this will pass.

But currently I do not really understand why local artifacts (as opposed to file:///tmp/n fake remote ones) are behaving differently.

Patch-set: 4
Label: Verified=0