Update patch set 2
Patch Set 2:

For this to work because GitCommitBear is unique and does not need a list of files but coala itself does. We need to add another plugin which accepts a list of files to parse.

As an example I've added the XMLBear and made it lint the settings.xml file. We can activate it for the rest of the XML as we fix XML linting (if that is something we want to pursue).

Another easy one could be to use the SpaceConsistencyBear to enforce indent style on files.

Patch-set: 2
Reviewer: Gerrit User 2759 <2759@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0