blueprint-maven-plugin file move made conditional
This makes the move of the generated blueprint XML file conditional,
because we should not unconditionally hard code the move of the
blueprint XML files generated by the blueprint-maven-plugin. 

This move from OSGI-INF/blueprint/autowire.xml to
org/opendaylight/blueprint should instead be conditional on the presence
of hand written XML in src/main/resources/org/opendaylight/blueprint
(which, typically, uses the ODL extensions).

This is because it should still be possible to use the @Singleton and
@Inject annotations with the XML generation by blueprint-maven-plugin in
bundles which do not depend on the MDSAL (e.g. in infrautils; I actually
have this case & problem in the real world).

Note that the gen. autowire.xml never uses the ODL extensions (unless we
do more work in the future to generate ODL extensions from annotations;
in which case the move should be made conditional on the XML content
actually containing/sing ODL extensions, not location).

Of course, normal ODL instance will always have the ODL controller mdsal
blueprint integration active.  But e.g. in ITs (e.g. in infrautils) it
seems "silly" to have to depend on the full mdsal JUST to get the
blueprint integration working JUST because we hard-coded and enforced
the move here.

This change has been tested by running a "mvn clean package" on e.g.
genius/interfacemanager/interfacemanager-impl, which is a typical ODL
bundle with a src/main/resources/org/opendaylight/blueprint/*.xml with
xmlns:odl="http://opendaylight.org/xmlns/blueprint/v1.0.0” as well as
some @Singleton in *.java, and verified that in the resulting JAR as
expected there are x2 XML in org/opendaylight/blueprint, namely the
handwritten say interfacemanager.xml as well as the generated
autowire.xml - both with and without this change; so no regressions.
Likewise, IFF there is no src/main/resources/org/opendaylight/blueprint,
then (only) the autowire.xml stays in OSGI-INF/blueprint (that's new,
and the goal of this change).

Change-Id: Id283078e0547e41caa0a714fe0f0fd70a048966c
Signed-off-by: Michael Vorburger <vorburger@redhat.com>