blueprint-maven-plugin gen. BP XML move made conditional
I HAVE NOT TESTED THIS AT ALL YET. How about self tests in odlparent? ;)

We should not unconditionally hard code the move of the blueprint XML
files generated by the blueprint-maven-plugin. Note that the gen. XML
never uses the ODL extensions (unless we do more work in the far future
to generate ODL extensions from annotations).

This move from OSGI-INF/blueprint/autowire.xml to
org/opendaylight/blueprint should isntead be conditional on the presence
of hand written XML in src/main/resources/org/opendaylight/blueprint
(which, typically, uses the ODL extensions).

This is because it should still be possible to use the @Singleton and
@Injec annotation with the XML generation by blueprint-maven-plugin in
bundles which do not depend on the MDSAL (e.g. in infrautils; I actually
have this case & problem in real world). 

Of course, normal ODL instance will always have the ODL controller mdsal
blueprint integration active.  But e.g. in ITs (e.g. in infrautils) it
seems "silly" to have to depend on the full mdsal JUST to get the
blueprint integration working JUST because we hard-coded and enforced
the move here.

Change-Id: Id283078e0547e41caa0a714fe0f0fd70a048966c
Signed-off-by: Michael Vorburger <vorburger@redhat.com>