pax-url-link & pax-exam-link-assembly in dependencyManagement
I'm experimenting with using http://veithen.github.io/alta/ to use Pax
URL's link: protocol and (eventually) completely do away with the mvn:
URLs from pax-url-aether (which is a stupid approach, not by us/ODL but
originally by Pax Exam).

Merging this isn't urgent yet, as I can just put the version into the
POM of where I'm experimenting with this, but ultimately it would be
good to have this in odlparent; just thought I'd raise it early.

Change-Id: Ic372ad5db0db9c26bbfce2f14592408365a1d5d7
Signed-off-by: Michael Vorburger <vorburger@redhat.com>