Relax Checkstyle RightCurly rule for LITERAL_WHILE
In the Checkstyle upgrade from 6.16 to 7.6.1 done in parallel in other
changes, I've come across a number of "violations" because it did not
like "... } while (condition)" in one line.

This indeed seems to be what our current Checkstyle rule would mandate,
but because of what looks like a bug in our currently used version 6.16
was never enforced.  As we upgrade, they seem to have fixed that, so
either we add a \n after } before while during the upgrade in the
projects which are already enforcing Checkstyle - or we just relax this
rule - personally I think it's ugly on two lines, and thus suggest this.

See http://checkstyle.sourceforge.net/config_blocks.html#RightCurly

I have not actually really tested this, just seems above right.

Change-Id: I21e20d91b854340a636c4fbbe5a6038caf0bee57
Signed-off-by: Michael Vorburger <vorburger@redhat.com>