Update patch set 5
Patch Set 5:

> > recheck
 > 
 > Ryan, the failure is caused by your commit message: you need to
 > reformat with 72-character lines at most for the body.

Sure.  I see that now though the failure cause was a bit embedded (I saw a failure to exec which is a bit misleading).  On this topic, that is an extremely anal requirement.  I get trying to make things look nice, but that seems a bit overly-stringent.  72 characters is not a lot...

Patch-set: 5
Reviewer: Gerrit User 3635 <3635@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0