Bug 7926: org.osgi.enterprise <scope>provided dependencyManagement
We suspect that the very confusing problems around Bug 7926 (causing
e.g. "java.net.BindException: Address already in use") had as a root
cause a bundle refresh of org.osgi.enterprise which broke down Pax Exam.

The c/55754 fixed the particular occurrence which we suspect behind
this, this change is complementary to that and should prevent similar
problems in the future.

Change-Id: I71356f48f4024934883b0c2364b9df5d0428cc91
Signed-off-by: Michael Vorburger <vorburger@redhat.com>