Log feature dependency changes
This enables the Karaf feature generator’s tracking features. When
features are generated, a simplified form is saved in
src/main/history; builds where this file is already present compare
newly-generated features against them, and log any failures. We don’t
fail builds where the features change (yet): any change to managed
dependencies would be liable to cause build failures in downstream
projects, requiring changes to those projects...

Any detected changes in features are logged as a warning, with the
detail of the changed feature contents. It would be interesting to
parse this in the builds and send emails when changes are detected.

Change-Id: If9a1ebb2b218413f78a1bc28bfeb20d39b956675
Signed-off-by: Stephen Kitt <skitt@redhat.com>