Allow logging feature dependency changes
This adds two properties controlling the Karaf feature generator’s
tracking features.

Setting checkDependencyChange to true enables checks and logs: when
features are generated, a simplified form  saved in src/main/history;
builds where this file is already present compare newly-generated
features against them, and log any failures. Any detected changes in
features are logged as a warning, with the detail of the changed
feature contents. It would be interesting to parse this in the builds
and send emails when changes are detected.

Setting failOnDependencyChange fails the build if a dependency change
is detected in the features.

This patch also enables both these flags in odlparent features where
it’s useful. This ensures that any changes to those features are
caught.

Change-Id: If9a1ebb2b218413f78a1bc28bfeb20d39b956675
Signed-off-by: Stephen Kitt <skitt@redhat.com>