Simplify feature dependencies
Our Karaf 4 features include a number of dependencies which are also
transitive dependencies, and which don’t constitute the primary
API/runtime we’re interested in providing. This patch removes these
features (the generated features are identical).

Once we remove the corresponding Karaf 3 features we’ll also be able
to remove some of these dependencies from dependency management
altogether.

Change-Id: Ifaa7db8afc1128c331b2cdb710dd95887c55788d
Signed-off-by: Stephen Kitt <skitt@redhat.com>