Introduce new aggregator-parent POM (and use it in odlparent itself)
Just to avoid having that <skip>true for maven-deploy-plugin and
maven-install-plugin copy/pasted all over ODL for each *-aggregator.

The intention is to start shifting to using this in other projects too.

Having a new parent artifact for all aggregators MAY also come in handy
when I start looking in earnest at our JavaDoc generation.

Q & TODO: Some, but not all, of the typical *-aggregator projects also
have a maven-site profile with maven-javadoc-plugin - should we move
that here as well? I'm hesitant because that sometimes has
<inherited>false .. which I'm not very familiar with - it would not do
anything with <inherited>false in a <parent>, would it? So do we include
it in aggregator-parent, but without <inherited>false ?  Or forget about
it and just start with a simple aggregator-parent as is proposed here
for now (and refine JavaDoc related stuff later) ?

Change-Id: I33fba54051b5d3e5fec145d8ae68785bb66c8725
Signed-off-by: Michael Vorburger <vorburger@redhat.com>