Bug 8415 related: TestBundleDiag API improved
required for https://git.opendaylight.org/gerrit/#/c/56749/

The implicit throws org.awaitility.core.ConditionTimeoutException and
java.lang.AssertionError were not great.. an AssertionError is special,
and more importantly Awaitility's ConditionTimeoutException is an
implementation detail.

Awaitility is actually bundled as a private OSGi embedded JAR-in-JAR,
which is a bit like shaded fat über JARs with relocation in non-OSGi
Java built by the maven-shade-plugin or maven-assembly-plugin; so it was
wrong to throw an exception from it to a consumer of this utility.

Therefore, this introduces a new SystemStateFailureException instead.

Change-Id: I7b48f5d39258a62c38a3fb8c98a6a35b97360276
Signed-off-by: Michael Vorburger <vorburger@redhat.com>