Bug 8415 related: Made TestBundleDiag an OSGi service (non-static)
required ASAP for infrautils.ready v2 which will work on Karaf 3 and 4.

includes "back-porting" changes made earlier to bundletest4 to the
original bundletest (what a PITA that we had to duplicate this code;
can't wait to remove all this once we're Karaf 4 only).

This will also come in handy if (when) the new ready service will
eventually have to be useable in non-OSGi environments such as component
tests (in which case an alternative implementation of the
API interface can be stubbed).

This also includes the new DiagUpdatesListener (to allow TestBundleDiag
to feedback async status to callers) which was originally reviewed under
c/56750.

Change-Id: I3f3a5037ea479b0fa8b8356b1470f4db595ccd89
Signed-off-by: Michael Vorburger <vorburger@redhat.com>