Bug 8415 related: TestBundleDiag as service (non-static)
required ASAP for infrautils.ready v2 which will work on Karaf 3 and 4.

This will also come in handy if (when) the new ready service will
eventually have to be useable in non-OSGi environments such as component
tests (in which case an alternative implementation of the
API interface can be stubbed).

Change-Id: I3f3a5037ea479b0fa8b8356b1470f4db595ccd89
Signed-off-by: Michael Vorburger <vorburger@redhat.com>