Update patch set 2
Patch Set 2:

Vratko, why this on stable/carbon, and only on stable/carbon?  I would have thought you would either want to make such a trivial change only on master (that would be my preference), or otherwise raise it for oth branches now (cherry-pick), to be sure it won't be forgotten?

Patch-set: 2