Bump org.apache.aries.blueprint.core version
From 1.4.2 to latest 1.8.1. because in
org.opendaylight.controller:blueprint version 1.7.1 was already used
anyway, and it seems wiser to declare this in only a single place, to
avoid possible version discrepancy problems re. this artifact in the
future.  So let's just upgrade to the very latest available here, and
drop the version in controller:blueprint in a follow-up change.

Change-Id: Ib332fb88eab11a376e93ea90304e2a22115b3e85
Signed-off-by: Michael Vorburger <vorburger@redhat.com>