Update patch set 1
Patch Set 1:

Not sure if this is the right way to do it but I searched for things that mentioned features4 and modified the skip variable. We really only want to disable features4 testing in Carbon so that we can turn on SFT in autorelease build again for karaf3 stuff.

Patch-set: 1