Update patch set 2
Patch Set 2:

if we want this to flow into 1.9.0 then we will need a bunch of downstream changes.  they essentially split shiro-core into four different artifacts, and the change is decidedly non-trivial.  id prefer we wait on merging this until we have a solid plan on how to adjust downstream consumers.

Patch-set: 2
Label: Code-Review=0