Update patch set 16
Patch Set 16:

The impacts are quite visible in the patches with the same topic. They boil down to Futures.transform() having been de-overloaded with transformAsync().

The yangtools patch actually addresses all the deprecation warnings except CheckedFuture.

I am not going to self-merge this and I doubt there will be time to spin a 3.0.0 in Nitrogen or willingness to push it through. That means this will be yet another thing on the backlog for Oxygen, which combined with scala and akka upgrades will be fun for whoever takes this on going forward.

Patch-set: 16
Reviewer: Gerrit User 649 <649@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0