Bug 8571: copy src/main/assembly
The contents of src/main/assembly need to be copied to
target/assembly, overwriting any conflicting file from Karaf.
integration/distribution needs this to be able to add custom scripts
and replace configuration files.

(Arguably this should be part of integration/distribution, but it
can’t be — at least not in a maintainable fashion.)

Change-Id: Ifcce82862d350b87853727e2c77140adee164886
Signed-off-by: Stephen Kitt <skitt@redhat.com>