Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

It mostly started as a way to stop projects from pushing their local karaf distributions to Nexus because they were supposed to be for local testing only and not release.

It's ok to release aggregator poms but the artifacts are kind of useless from an end user perspective as they are only useful to run a build so the idea was to not release them but there's really no harm in releasing them.

Patch-set: 1
Label: Code-Review=-1