Update patch set 1
Patch Set 1: Code-Review+2

I tested this patch as best I could with the bad pax-web-feature file in place. As far as I can tell it works around the issue we are seeing when controller/bgpcep tries to build their own local karafs and failing to patch. I tried to do a comparison between the files produced by this patch and without this patch and as far as I can tell they appear to be consistent.

Robert, if you agree I think we should merge this and release a odlparent 2.0.1 release and see if it fixes the issues with the distribution-karaf builds in controller and bgpcep.

Patch-set: 1
Reviewer: Gerrit User 2759 <2759@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+2