${project.build.directory} in maven-bundle-plugin
Instead of hard-coding "target/" as directory name.
for the LICENSE file. This fixes the following probem:

   mvn -Pq,ide clean install

[INFO] --- maven-bundle-plugin:3.0.1:bundle (default-bundle) @ ... ---
[ERROR] Bundle ... : Input file does not exist: target/classes/LICENSE
[ERROR] Error(s) found in bundle configuration

Locally verified that with this fix, the problem shown above does not
happen anymore - and the LICENSE file is still in bundle JARs.

PS: In order to reproduce the problem above (and test this fix),
you (obviously) need to 'rm -rf target/ target-ide/' first... ;)

Change-Id: I313478ec95c9a6e647a595418bd92d45bbb04a8d
Signed-off-by: Michael Vorburger <vorburger@redhat.com>