Add Javadoc links for third parties
Add some links for third party libraries declared as dependencies in
odlparent-lite so that those links are automatically added to the Maven
javadoc:aggregate or javadoc:javadoc goals. The list is not
comprehensive, it is just an starting point to continue adding more
links.

The projects/modules/aggregators referencing odlparent-lite will inherit
these links.

The <detectLinks> flag is set to false until [1] is resolved.

[1] https://issues.apache.org/jira/browse/MJAVADOC-427

Change-Id: I56a6eabe0ba7b9f32ffed7748ab13a79aa6bc2e1
Signed-off-by: David Suarez <david.suarez.fuentes@gmail.com>