Use log4j instead of slf4j-simple for tests
Required so that unit tests can use the the MDC, and that we can write
unit tests of utilities related to this, see the related work ongoing
under https://git.opendaylight.org/gerrit/#/q/topic:mdc.

Even if strictly speaking not everyone may be interested in using
logging with MDC, the problem is that because we (OK, me..) made
slf4j-simple a default dependency for EVERYTHING in odlparent, no child
project can now "opt out" and use another implementation anymore now, as
there is no <excludes> for something inherited from a parent POM in
Maven.  Therefore, this proposes to simply change it.

The full log4j takes a few more cycles to initialize than slf4j-simple,
so this does make all unit tests a tiny bit slower to start up - but
IMHO that is acceptable.  (If someone can think of a better way with
Maven to have a default logger like slf4j-simple, but at the same time
let some projects opt to choose another one like log4j, I'd be
interested to see counter proposals to this.)

https://git.opendaylight.org/gerrit/#/c/61521/ adjusts infrautils' use.

Change-Id: I665639512a3530cb87915ad20676605a65a9f0ac
Signed-off-by: Michael Vorburger <vorburger@redhat.com>