Bug 8993: Patch karaf paxweb to bump beanutils
from commons-beanutils/1.8.3 to commons-beanutils/1.9.3

due to (apparently) security violation concerns in 1.8.3

This change actually has no effect, and does not really
change the version of beanutils in karaf's system/ because
it only touches paxweb's pax-jsf-support feature, which we
do not use; but it seems to me to still be a good idea to do.

I've not been able to figure out what else drags along 1.8.3.

Change-Id: I644e3e830f5dae437ae1f10ae9e14d2c653568c2
Signed-off-by: Michael Vorburger <vorburger@redhat.com>