Update patch set 1
Patch Set 1:

also -1 can we please have that error_prone_annotations JAR as a default dependency everywhere, just like we do for FindBugs' and the Eclipse NonNull ones.  It causes no harm (perhaps see if we could even make it <scope>provided?) and just makes it easier to use them easily, instead of have to edit each pom.xml to add it.

Patch-set: 1