Bump jdt.annotation from 2.0.0 to 2.1.0
I'm starting to explore using org.eclipse.jdt.annotation more, and want
to have its latest version at my disposal for this.  This is part of
https://git.opendaylight.org/gerrit/#/c/47370/ but that's bigger and
still WIP and it would be great if you would agree to get this one in
already now.

This should be very low risk and not break anything anywhere.

Change-Id: Ib1f6389cc0929e08d1585efc7bdc5e75097a3544
Signed-off-by: Michael Vorburger <vorburger@redhat.com>