Bug 9169: Bump jdt.annotation from 2.0.0 to 2.1.0
Most importantly, this introduces a handy util for legacy interop
org.eclipse.jdt.annotation.Checks; see its JavaDoc on:
https://help.eclipse.org/neon/index.jsp?topic=%2Forg.eclipse.jdt.doc.isv%2Freference%2Fapi%2Forg%2Feclipse%2Fjdt%2Fannotation%2FChecks.html

Full git change log with more details available on
https://gist.github.com/vorburger/0479be76f9cd3d10aa5ef86d4034a090

I'm starting to explore using org.eclipse.jdt.annotation more, and want
to have its latest version at my disposal for this.  This is part of
https://git.opendaylight.org/gerrit/#/c/47370/ but that's bigger and
still WIP and it would be great if you would agree to get this one in
already now.

This should be very low risk and not break anything anywhere.

Change-Id: Ib1f6389cc0929e08d1585efc7bdc5e75097a3544
Signed-off-by: Michael Vorburger <vorburger@redhat.com>