Update patch set 2
Patch Set 2: -Code-Review

> Hang on, before merging this, lemme figure out how to correctly
 > include this JAR into a feature so that SFT does not fail as

no, removing self -2, I figured that I do WANT this in, please; cauz: That failure is BECAUSE of this missing - SFT passes if you work around that as done in https://git.opendaylight.org/gerrit/#/c/63963/3..4/common/features/odl-infrautils-utils/pom.xml ... but we shouldn't have to do that, of course.

PS: Karaf is not such a PITA after all... ;)  BTW: Thank you Karaf Maven Plugin for auto-generating the features.xml for me - so that it does "just work"!

Patch-set: 2