Update patch set 5
Patch Set 5:

> This would mean that all downstreams need to add the appropriate
 > SPDX headers to all their Java files (at least) as part of their
 > odlparent 3 upgrade, right? If the medium-term plan is to switch to
 > a license checking job, wouldn’t it be better to do that instead?
 > It seems to me we’re adding a huge amount of work (admittedly,
 > mechanisable work) to the odlparent 3 upgrade, where we could
 > instead roll it out progressively using a separate job...

Right there's going to be a transition period either way and I'm willing to submit patches to fix projects. You might be right though we could push the job and then do the job + fixing spdx headers as part of the same task. Only thing is if we go the job route we also need to disable the license checker in odlparent otherwise it will fail builds once we add spdx headers.

Ok. Let me push on getting the job ready and then I can do a big push to get projects to merge license header updates all at once along with the job. Then we can change this patch to instead disable the checkstyle license checker.

Patch-set: 5