Update patch set 5
Patch Set 5:

Sort of. My plan is to have a single job (not one for every project) that will trigger against Gerrit patch upload to any project and run the checker (This is to ensure that no project can bypass the checker or we forget to add the job to any project). We can definitely do it in phases though so this is what I'm thinking now:

1) Deploy license checker job without spdx verification
2) Disable odlparent checkstyle license checker
3) Submit patches to all projects to add spdx headers
4) Enable license checker job with spdx verification

At this point projects must merge their patches (and their jobs will fail until they do) but reactive projects will merge patches quickly and projects that are slow will only be blocking themselves.

Since we have a script to insert the headers to the java files it's not too bad workload wise.

Patch-set: 5