Update patch set 2
Patch Set 2:

The reason I have a problem with scope=compile is that it adds dependency bloat to everyone, everywhere, in production artifacts. Vast majority of our components do not need this, as they are just using JDK/Guava equivalents.

If a component needs Checks, it should override the scope.

Patch-set: 2