Update patch set 10
Patch Set 10:

> How do more complex projects react to this?

I'd love to see people try it, and provide feedback.  I could refine some of the configuration here, if there are shared issues.  My starting point is that even in complex projects, there's no real good reason that I can think of for 1 artifact's classpath to have duplicate classes - if that is the case, it's a risk for that project's build (and runtime) stability, because they can't be 100% sure which of their conflicting dependencies the build (and run) against.

Patch-set: 10
Label: Verified=0