Add missing M2E lifecycle mappings metadata
* Ignore for the karaf-maven-plugin

This change does intentionally NOT include a lifecycle mapping for the
(new) spotbugs-maven-plugin, because that's one which we do NOT want to
just IGNORE, but have to configure; re. that one, please watch
https://github.com/m2e-code-quality/m2e-code-quality/issues/110

FTR: Originally I started putting these (only) into
opendaylight-eclipse-setup, because "such Eclipse specifics should not
be in our POM", but today on IRC rovarga opined that "it should 'just
work' even if one doesn't want to use your opendaylight-eclipse-setup"
but why? its great!] so putting there here now - meaning we have a bit
of a mix now, but I guess this isn't the end of the world.

see https://github.com/vorburger/opendaylight-eclipse-setup#how-to-add-additional-m2e-lifecycle-mapping

Change-Id: I3498cc5f3af7481cd83b4f7f69045d19668fd1da
Signed-off-by: Michael Vorburger <vorburger@redhat.com>