Update patch set 2
Patch Set 2:

> Why do you want this in 2.0.x?

because I want to move forward with INFRAUTILS-19 before the end of year after next, and am not clear if we're on track with 3.0.0 so am thinking if not we should have a 2.0.6 as a fallback plan - what do you think?  We're clearly actively maintaing 2.0.x, given https://git.opendaylight.org/gerrit/#/c/65942/ and https://git.opendaylight.org/gerrit/#/c/65932/, so why not this also?  (As well as https://git.opendaylight.org/gerrit/#/c/66030/ and its two parents.)

Patch-set: 2