Update patch set 1
Patch Set 1:

This is a debating point, but I prefer getClass() because it fails on the safe side when changes are made without considering the ramifications. https://stackoverflow.com/q/596462/4500798 is the best reference currently (and the resources it points to).

Patch-set: 1
Reviewer: Gerrit User 4540 <4540@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0