Update patch set 3
Patch Set 3:

I am sorry, but users are confused all the time.

What you are implying is that odlparent should actively stand in the way of downstreams allowing to use whatever works in the name of users not shooting themselves in the foot.

Furthermore, you are implying that there needs to be an explicit odlparent release which downstreams need to adopt for this hurdle to go away.

No, odlparent certainly should not be unilaterally making these sorts of policy decisions.

Patch-set: 3