Bump spotbugs-maven-plugin from 3.1.0-RC6 to 3.1.1
This requires a newer Maven version than 3.3.9, at the very least in
order to be able to "mvn deploy" without hitting a weird internal
Maven error; therefore this change also (kind of has to....) bump the
requireMavenVersion in the maven-enforcer-plugin in odlparent-lite.

The jump from 3.3.9 to 3.5.2 is just because that's the latest one.
(Robert confirms even 3.5.0 would be fine; we don't know about 3.4.x.)

As the opendaylight.org Jenkins Build Servers are still on mvn 3.3.9,
this change will have to wait for the mvn version on those to be
bumped before this can be merged; this has been proposed in helpdesk
Ticket #50800, and could happen e.g. at the start of our next major
release Fluorine.

see extensive discussion on email lists thread
"Failing verify jobs due to weird low level Maven error during deploy"
https://lists.opendaylight.org/pipermail/odlparent-dev/2018-January/thread.html

Change-Id: I1325e628ce62bd0f8d426d7cc61246003fc06aed
Signed-off-by: Michael Vorburger <vorburger@redhat.com>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>