Update patch set 3
Patch Set 3:

Except this is targeted at 3.0.x, which has zero issues with java 9, as far as I can tell.

Furthermore, if javadocs are failing to build, we should introduce a profile which either disables them completely or adds whatever knob is needed to make them work with jdk9.

We've done exactly that during jdk8 migration, I do not see why it is not feasible now -- as far as I can tell, that was not even considered.

Patch-set: 3